dsmr/requirements.txt
pyserial
paho-mqtt
influxdb
